Offer the unminified package by default
While the minified package is useful for developers who just want to get started quickly and download the package, I’d imagine that most people pulling the package in via NPM or equivalent will have their own minification setup. In this case it doesn’t save time and makes it harder to debug the library.